libumidi: Fix compile warnings.

+++ umidi20_coremidi.c
@@ -339,7 +339,7 @@ umidi20_coremidi_uniq_inputs(char **ptr)
 	for (x = 0; x != n; x++) {
 		for (z = 0, y = x + 1; y != n; y++) {
 			if (strcmp(ptr[x], ptr[y]) == 0) {
-				int s = (int)(strlen(ptr[y]) + 16);
+				size_t s = strlen(ptr[y]) + 16;
 				pstr = ptr[y];
 				ptr[y] = malloc(s);
 				if (ptr[y] == NULL)
